+++ src/components/core/navigation/index.ts
@@ -1,2 +1,3 @@
 export { default as Link } from './Link';
 export { default as SideBar } from './SideBar';
+export { default as TabBar } from './TabBar';